requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Werkzeug==3.1.3
requests==2.31.0
numpy==2.3.1
joblib==1.4.2
pytz==2024.1
pytest==8.4.1
pytest-cov==6.2.1
pytest-flask==1.3.0
python-dotenv==1.0.1
pyotp==2.9.0
qrcode==7.4.2
Pillow==10.4.0
PyJWT==2.8.0 